+++ util/failure/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Changed
+- Add optional `variant` attribute to `SerializableFail`.
 - Serialisation of `&Fail` instead of `Fail`.
 
 ## 0.1.1 - 2019-06-16